2010-05-27	Ralf Corsépius <ralf.corsepius@rtems.org>
	*  libnetworking/lib/ftpfs.c, libnetworking/lib/tftpDriver.c:
	Use size_t instead of int for pathnamelen.

+++ cpukit/libnetworking/lib/ftpfs.c
@@ -1196,7 +1196,7 @@ static int rtems_ftpfs_ftruncate(rtems_libio_t *iop, rtems_off64_t count)
 
 static int rtems_ftpfs_eval_path(
   const char *pathname,
-  int pathnamelen,
+  size_t pathnamelen,
   int flags,
   rtems_filesystem_location_info_t *pathloc
 )
