+++ cpukit/libnetworking/lib/tftpDriver.c
@@ -488,7 +488,7 @@ fixPath (char *path)
 
 static int rtems_tftp_eval_path(
   const char                        *pathname,     /* IN     */
-  int                                pathnamelen,  /* IN     */		
+  size_t                             pathnamelen,  /* IN     */		
   int                                flags,        /* IN     */
   rtems_filesystem_location_info_t  *pathloc       /* IN/OUT */
 )
